Adding some delete route edits - does not work

+++ app.py
@@ -7,6 +7,7 @@
 #           | |
 #           |_|
 import os
+from re import I
 from dotenv import load_dotenv
 from flask import Flask, redirect, render_template, request, abort, url_for, session, url_for, flash, send_from_directory
 from sqlalchemy import null
@@ -63,7 +64,7 @@ load_dotenv()
 db_host = os.getenv('DB_HOST', 'localhost')
 db_port = os.getenv('DB_PORT', '3306')
 db_user = os.getenv('DB_USER', 'root')
-db_pass = os.getenv('DB_PASSWORD',)
+db_pass = os.getenv('DB_PASSWORD', '8569742Dl!')
 db_name = os.getenv('DB_NAME', 'picnetic_db')
 
 connection_string = f'mysql://{db_user}:{db_pass}@{db_host}:{db_port}/{db_name}'
@@ -226,10 +227,10 @@ def userAccount(user_id):
         comment_his = userAccount.comments
         post_his = userAccount.posts
         print(userAccount)
-        if comment_his or post_his is None:
+        if comment_his and post_his is None:
             return render_template("account2.html", userAccount=userAccount)
 
-    return render_template("account.html", userAccount=userAccount, comment_his=comment_his, post_his=post_his, ui = userid)
+    return render_template("account.html", userAccount=userAccount, comment_his=comment_his, post_his=post_his, ui = session['user']['user_id'])
 
 
 def is_allowed(filename):
@@ -418,25 +419,24 @@ def not_found(e):
     return render_template('404.html'), 404
 
 
-@app.route('/delete/<user_id>', methods=['GET'])
+@app.route('/delete/<user_id>', methods=['POST', 'GET'])
 def delete(user_id):
     if 'user' in session:
         userid = session['user']['user_id']
         userAccount = Users.query.get(userid)
-
-        # delete users comments
-        comment_his = userAccount.comments
-        db.session.delete(comment_his)
-        # delete users posts
-        post_his = userAccount.posts
-        db.session.delete(post_his)     
-        # finally, delete the user's account
-        db.session.delete(userAccount)
-        db.session.commit()
-
-        flash("Your account has been deleted!")
-
-        return redirect(url_for('index'))
+        if request.method == 'POST':
+            # delete users comments
+            comment_his = userAccount.comments
+            for c in comment_his:
+                db.session.delete(c)
+            # delete users posts
+            post_his = userAccount.posts
+            for p in post_his:
+                db.session.delete(p)    
+            # finally, delete the user's account
+            db.session.delete(userAccount)
+            db.session.commit()
+    return render_template('deletedAccount.html')
 
 
 if __name__ == '__main__':
